+++ unionfind/src/main/java/actors/CorrelatorActor.java
@@ -3,7 +3,6 @@ package actors;
 import java.util.List;
 import akka.actor.AbstractActor;
 import akka.actor.ActorRef;
-import algorithm.Event;
 import algorithm.ICorrelator;
 import tree_union_find.IdCorrelatorTree;
