+++ unionfind/src/main/java/actors/DeleteGroupRequest.java
@@ -1,7 +1,5 @@
 package actors;
 
-import algorithm.Event;
-
 public class DeleteGroupRequest {
 	private Event event;
