+++ unionfind/src/main/java/actors/UpdateStatusRequest.java
@@ -2,8 +2,6 @@ package actors;
 
 import java.util.List;
 
-import algorithm.Event;
-
 public class UpdateStatusRequest {
 	private final List<Event>  eventList;
 	public UpdateStatusRequest(List<Event>  events) {
