+++ unionfind/src/test/java/algorithm/CorrelatorActorTest.java
@@ -5,6 +5,7 @@ import java.util.ArrayList;
 import org.junit.jupiter.api.Test;
 
 import actors.DeleteGroupRequest;
+import actors.Event;
 import actors.IdCorrelatorManager;
 import actors.TracingActorSystem;
 import akka.actor.ActorRef;
